Remove deprecated status method
